Moved parse_masterkey_file back to triage/masterkeys.py
