+++ .gitignore
@@ -1,4 +1,4 @@
-venv
+*venv
 *.egg-info
 build
 dist
